refactor: update import paths for my-filters components and introduce new components
- Refactored import statements for various components within the my-filters feature, moving them to a new `features/my-filters` directory for improved organization.
- Introduced new components including `FilterImageCombobox`, `FilterImage`, `MyFilterCard`, and `MyFilters`, enhancing the functionality and structure of the my-filters feature.
- Updated relevant files to reflect these changes, improving code clarity and maintainability.
