+++ src/app/(app)/(stacked-layout)/my-filters/new-filter/new-filter-form.tsx
@@ -26,7 +26,7 @@ import {
 import { Input } from "@/components/ui/input";
 import { ConveyorCard } from "@/components/features/conveyor/conveyor-card";
 import { FilterCategoryCombobox } from "@/components/features/my-filters/components/filter-category-combobox";
-import { FilterImageCombobox } from "@/components/filter-image-combobox";
+import { FilterImageCombobox } from "@/components/features/my-filters/components/filter-image-combobox";
 
 const DevTool = dynamic(
   () => import("@hookform/devtools").then((module) => module.DevTool),
