+++ src/components/features/my-filters/components/filter-image.tsx
